% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArgTiles.R
\name{addArgTiles}
\alias{addArgTiles}
\title{Agrega capa de Argentina para mapa interactivo
(\emph{Load Argentina Tiles Template})}
\usage{
addArgTiles(data)
}
\arguments{
\item{data}{debe ser un objeto con class "leaflet" "htmlwidget".}
}
\description{
Función que descarga capa base de Argentina para mapa interactivo creado con la libreria \code{leaflet}.
}
\details{
El objetivo es agregar referencias de la base sobre la que se visualizan geometrías de Argentina. Se pueden agregar capas como polígonos descargandolos con \code{\link{get_geo}}.
Los geometrías disponibles se pueden chequear con \code{\link{show_arg_codes}}.
}
\examples{

 get_geo("TUCUMAN") \%>\%
 leaflet::leaflet() \%>\%
 leaflet::addPolygons() \%>\%
 addArgTiles()


}
