% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censos.R
\docType{data}
\name{censos}
\alias{censos}
\title{Un archivo de datos que contiene el lsitado de las geometrías para los censos disponibles}
\format{
Un tibble con 526 filas y 8 variables:
\describe{
\item{ceso}{ID correspondiente al año del censo de Argentina para descargar poligonos disponibles

}}
}
\usage{
censos
}
\description{
Un archivo de datos que contiene el lsitado de las geometrías para los censos disponibles
}
\keyword{datasets}
