% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_metadata.R
\docType{data}
\name{geo_metadata}
\alias{geo_metadata}
\title{Un archivo de datos que contiene identificadores geográficos para los departamentos de Argentina y sus provincias
(\emph{A data file containing geographic identifiers for the departments of Argentina and their provinces})}
\format{
Un tibble con 526 filas y 8 variables:
\describe{
\item{codprov, coddepto}{identificadores utilizados por \emph{INDRA} para escrutinios provisorios de elecciones
nacionales}
\item{\emph{x_censo}}{identificadores de unidades geográficas del \emph{INDEC}}
\item{\emph{x_iso}}{identificadores de \emph{International Organization for Standardization}}

...
}
}
\usage{
geo_metadata
}
\description{
Un archivo de datos que contiene identificadores geográficos para los departamentos de Argentina y sus provincias
(\emph{A data file containing geographic identifiers for the departments of Argentina and their provinces})
}
\keyword{datasets}
