% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bahra.R
\name{get_bahra}
\alias{get_bahra}
\title{Descarga 'base total' de Asentamientos Humanos de la República Argentina (BAHRA)}
\usage{
get_bahra(geo = "ARGENTINA")
}
\arguments{
\item{geo}{un character con el nombre del distrito que se quiere descargar (por defecto toda ARGENTINA)
Se puede chequear el id con \code{\link{show_arg_codes}}.}
}
\value{
tibble con datos de BAHRA
}
\description{
Descarga 'base total' de Asentamientos Humanos de la República Argentina (BAHRA)
}
\examples{
get_bahra()
}
