% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_censo.R
\name{get_censo}
\alias{get_censo}
\title{Carga poligonos de Censos de Población Históricos de Argentina}
\usage{
get_censo(censo = NULL, simplified = FALSE)
}
\arguments{
\item{censo}{un character con el id del año del censo del district que se quiere descargar.
Se pueden chequear el id con \code{\link{show_available_censos}}.}

\item{simplified}{por defecto es TRUE y determina la descarga de una versión simplificada de las geometrias.
Con FALSE descarga la versión original de INDEC}
}
\value{
tibble con capa geografica correspondiente a geometrías de un censo histórico
}
\description{
Carga poligonos de Censos de Población Históricos de Argentina
}
\examples{
get_censo(censo = "1947", simplified = TRUE)

}
