% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eph.R
\name{get_eph}
\alias{get_eph}
\title{Carga poligonos de los Aglomerados Urbanos correspondientes a la Encuesta Permanente de Hogares (INDEC)}
\usage{
get_eph(geo = "ARGENTINA", simplified = TRUE)
}
\arguments{
\item{geo}{un character con el nombre del district que se quiere descargar.
Se pueden chequear el id con \code{\link{show_arg_codes}}.}

\item{simplified}{por defecto es TRUE y determina la descarga de una versión simplificada de las geometrias.
Con FALSE descarga la versión original de INDEC}
}
\description{
Carga poligonos de los Aglomerados Urbanos correspondientes a la Encuesta Permanente de Hogares (INDEC)
}
\examples{
get_eph(geo = "TUCUMAN")

}
