% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo.R
\name{get_geo}
\alias{get_geo}
\title{Carga poligonos geográficos de distritos de Argentina
(\emph{Load Argentina's districts geometries})}
\usage{
get_geo(geo = NULL, level = "departamento")
}
\arguments{
\item{geo}{un character con el nombre del district que se quiere descargar. Se pueden chequear el id con \code{\link{show_arg_codes}}.}

\item{level}{parametro opcional para descargar geometrías a nivel 'departamento' cuando se solicita mapa nacional \code{get_geo(geo = "ARGNTINA", level = "departamento")}.}
}
\description{
Función que descarga (\emph{geometry}) para graficar con mapas
}
\examples{

get_geo("TUCUMAN")

}
