% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefar.R
\name{get_localidades}
\alias{get_localidades}
\title{Obtener Localidades}
\usage{
get_localidades(
  id = NULL,
  nombre = NULL,
  provincia = NULL,
  departamento = NULL,
  municipio = NULL,
  interseccion = NULL,
  orden = NULL,
  aplanar = TRUE,
  campos = NULL,
  max = NULL,
  inicio = NULL,
  exacto = NULL
)
}
\arguments{
\item{id}{text Filtrar por ID.}

\item{nombre}{text Filtrar por Nombre.}

\item{provincia}{text Filtrar por nombre o ID de Provincia.}

\item{departamento}{text Filtrar por nombre o ID de Departamento.}

\item{municipio}{text Filtrar por nombre o ID de Municipio.}

\item{interseccion}{text Geometría GeoJSON utilizada para filtrar resultados por intersección espacial. Sólo se soportan polígonos y multipolígonos. Ejemplo: polygon((-58.431,-34.592),(-58.430,-34.590),(-58.428,-34.593),(-58.431,-34.592)).}

\item{orden}{text Campo por el cual ordenar los resultados (por ID o nombre)}

\item{aplanar}{boolean Cuando está presente, muestra el resultado JSON con una estructura plana.}

\item{campos}{text Campos a incluir en la respuesta separados por comas, sin espacios. Algunos campos siempre serán incluidos, incluso si no se agregaron en la lista. Para incluir campos de sub-entidades, separar los nombres con un punto, por ejemplo: provincia.id.}

\item{max}{integer Cantidad máxima de resultados a devolver. La API limita a un máximo de 2000 para este endpoint.}

\item{inicio}{integer Cantidad de resultados a omitir desde el principio.}

\item{exacto}{boolean Cuando está presente, se activa el modo de búsqueda por texto exacto. Sólo tiene efecto cuando se usan campos de búsqueda por texto (por ejemplo, nombre).}
}
\value{
Un Data Frame con el listado de Localidades
}
\description{
Permite realizar búsquedas sobre el listado de localidades.
Realiza la consulta GET al endpoint /localidades de georef-ar-api.
Si existe GEOREFAR_TOKEN en el Renviron lo usará para hacer la consulta.
}
\examples{
\dontrun{
get_localidades(nombre = "PALERMO", provincia = "CIUDAD AUTONOMA DE BUENOS AIRES")
}
}
\references{
\href{https://datosgobar.github.io/georef-ar-api/open-api/#/Recursos/get_localidades}{georef-ar-api/localidades}
}
