% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefar.R
\name{normalizar_direccion}
\alias{normalizar_direccion}
\title{Normalizacion de direcciones}
\usage{
normalizar_direccion(
  direccion,
  tipo = NULL,
  provincia = NULL,
  departamento = NULL,
  aplanar = TRUE,
  campos = NULL,
  max = NULL,
  exacto = NULL
)
}
\arguments{
\item{direccion}{text Requerido. Direccion a normalizar, debe contener altura separada por espacio. (Ej: Colon 127)}

\item{tipo}{text Tipo de calle. (Valores disponibles: calle, avenida, pasaje.)}

\item{provincia}{text Filtrar por nombre o ID de provincia.}

\item{departamento}{text Filtrar por nombre o ID de departamento.}

\item{aplanar}{boolean Cuando está presente, muestra el resultado JSON con una estructura plana.}

\item{campos}{text Campos a incluir en la respuesta separados por comas, sin espacios. Algunos campos siempre serán incluidos, incluso si no se agregaron en la lista. Para incluir campos de sub-entidades, separar los nombres con un punto, por ejemplo: provincia.id.}

\item{max}{integer Cantidad máxima de resultados a devolver.}

\item{exacto}{boolean Cuando está presente, se activa el modo de búsqueda por texto exacto. Sólo tiene efecto cuando se usan campos de búsqueda por texto (por ejemplo, nombre).}
}
\value{
Un Data Frame con el listado normalizado de de direcciones
}
\description{
Normalizacion de direcciones
}
\examples{
\dontrun{
normalizar_direccion()
}
}
