% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefar.R
\name{post_asentamientos_bulk}
\alias{post_asentamientos_bulk}
\title{Enviar Lote de Consultas de Asentamientos (POST)}
\usage{
post_asentamientos_bulk(queries_list)
}
\arguments{
\item{queries_list}{Lista de listas. Cada lista interna debe contener los parámetros
para una consulta de asentamiento individual.
Parámetros válidos por consulta: id, nombre, provincia, departamento, municipio, localidad_censal, orden, aplanar, campos, max, exacto.}
}
\value{
Un Data Frame (tibble) con los resultados combinados de todas las consultas.
}
\description{
Permite realizar múltiples búsquedas sobre el listado de asentamientos BAHRA en una sola llamada POST.
Realiza la consulta POST al endpoint /asentamientos de georef-ar-api.
}
\examples{
\dontrun{
consultas_asent <- list(
  list(provincia = "CHUBUT", nombre = "COSTA")
)
resultados_asent <- post_asentamientos_bulk(queries_list = consultas_asent)
print(resultados_asent)
}
}
\references{
\href{https://datosgobar.github.io/georef-ar-api/open-api/#/Recursos/post_asentamientos}{georef-ar-api/asentamientos POST}
}
