% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censos.R
\docType{data}
\name{censos}
\alias{censos}
\title{Un archivo de datos que contiene el listado de las geometrías para los censos disponibles}
\format{
Un tibble con 10 filas filas y 1 variables (censo)que funciona como id para usar con \code{\link{get_geo}}
\describe{
\item{censo}{ID correspondiente al año del censo de Argentina para descargar poligonos disponibles (1869 - 2010)

}}
}
\usage{
censos
}
\description{
Un archivo de datos que contiene el listado de las geometrías para los censos disponibles
}
\keyword{datasets}
