% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoAr.R
\docType{package}
\name{geoAr}
\alias{geoAr}
\title{\code{geoAr} package}
\description{
Caja de Herramientas GEO de Argentina
See the README on
\href{https://github.com/PoliticaArgentina/geoAr/README.md}{Github}
}
