% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefar.R
\name{get_localidades}
\alias{get_localidades}
\title{Obtener Localidades}
\usage{
get_localidades(
  id = NULL,
  nombre = NULL,
  provincia = NULL,
  departamento = NULL,
  municipio = NULL,
  orden = NULL,
  aplanar = TRUE,
  campos = NULL,
  max = NULL,
  exacto = NULL
)
}
\arguments{
\item{id}{text Filtrar por ID.}

\item{nombre}{text Filtrar por Nombre.}

\item{provincia}{text Filtrar por nombre o ID de Provincia.}

\item{departamento}{text Filtrar por nombre o ID de Departamento.}

\item{municipio}{text Filtrar por nombre o ID de Municipio.}

\item{orden}{text Campo por el cual ordenar los resultados (por ID o nombre)}

\item{aplanar}{boolean Cuando está presente, muestra el resultado JSON con una estructura plana.}

\item{campos}{text Campos a incluir en la respuesta separados por comas, sin espacios. Algunos campos siempre serán incluidos, incluso si no se agregaron en la lista. Para incluir campos de sub-entidades, separar los nombres con un punto, por ejemplo: provincia.id.}

\item{max}{integer Cantidad máxima de resultados a devolver.}

\item{exacto}{boolean Cuando está presente, se activa el modo de búsqueda por texto exacto. Sólo tiene efecto cuando se usan campos de búsqueda por texto (por ejemplo, nombre).}
}
\value{
Un Data Frame con el listado de Departamentos
}
\description{
Obtener Localidades
}
\examples{
\dontrun{
get_localidades()
}
}
