% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_censos.R
\name{show_available_censos}
\alias{show_available_censos}
\title{Geometrías de CENSOS
(\emph{geo ID's dictoniary})}
\usage{
show_available_censos()
}
\description{
Función que devuelve un \emph{data.frame} con listado de las geometrías de los CENSOS disponible
}
\examples{

show_available_censos()

}
